+++ src/main/java/kr/hhplus/be/domain/coupon/entity/Coupon.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.domain.coupon.entity;
 import jakarta.persistence.*;
 import kr.hhplus.be.domain.coupon.enumtype.CouponStatus;
 import kr.hhplus.be.domain.coupon.enumtype.DiscountType;
-import kr.hhplus.be.support.exception.CommerceConflictException;
-import kr.hhplus.be.support.exception.ErrorCode;
+import kr.hhplus.be.domain.exception.CommerceConflictException;
+import kr.hhplus.be.domain.exception.ErrorCode;
 import lombok.*;
 
 import java.time.LocalDateTime;
