+++ src/main/java/kr/hhplus/be/infrastructure/coupon/repository/CouponPublishRepositoryImpl.java
@@ -5,8 +5,8 @@ import kr.hhplus.be.domain.coupon.entity.CouponPublish;
 import kr.hhplus.be.domain.coupon.repository.CouponPublishRepository;
 import kr.hhplus.be.domain.user.dto.UserCouponDTO;
 import kr.hhplus.be.infrastructure.coupon.jpa.CouponPublishJpaRepository;
-import kr.hhplus.be.support.exception.CommerceNotFoundException;
-import kr.hhplus.be.support.exception.ErrorCode;
+import kr.hhplus.be.domain.exception.CommerceNotFoundException;
+import kr.hhplus.be.domain.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
