+++ src/main/java/kr/hhplus/be/infrastructure/product/repository/ProductInventoryRepositoryImpl.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.infrastructure.product.repository;
 import kr.hhplus.be.domain.product.entity.ProductInventory;
 import kr.hhplus.be.domain.product.repository.ProductInventoryRepository;
 import kr.hhplus.be.infrastructure.product.jpa.ProductInventoryJpaRepository;
-import kr.hhplus.be.support.exception.CommerceNotFoundException;
-import kr.hhplus.be.support.exception.ErrorCode;
+import kr.hhplus.be.domain.exception.CommerceNotFoundException;
+import kr.hhplus.be.domain.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
