+++ src/test/java/kr/hhplus/be/domain/coupon/entity/CouponTest.java
@@ -2,8 +2,8 @@ package kr.hhplus.be.domain.coupon.entity;
 
 import kr.hhplus.be.domain.coupon.enumtype.CouponStatus;
 import kr.hhplus.be.domain.coupon.enumtype.DiscountType;
-import kr.hhplus.be.support.exception.CommerceConflictException;
-import kr.hhplus.be.support.exception.ErrorCode;
+import kr.hhplus.be.domain.exception.CommerceConflictException;
+import kr.hhplus.be.domain.exception.ErrorCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
