+++ src/test/java/kr/hhplus/be/domain/user/entity/UserTest.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.domain.user.entity;
 
-import kr.hhplus.be.support.exception.CommerceConflictException;
-import kr.hhplus.be.support.exception.ErrorCode;
+import kr.hhplus.be.domain.exception.CommerceConflictException;
+import kr.hhplus.be.domain.exception.ErrorCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
