+++ src/test/java/kr/hhplus/be/domain/user/service/UserServiceTest.java
@@ -9,8 +9,8 @@ import kr.hhplus.be.domain.user.entity.BalanceHistory;
 import kr.hhplus.be.domain.user.entity.User;
 import kr.hhplus.be.domain.user.repository.BalanceHistoryRepository;
 import kr.hhplus.be.domain.user.repository.UserRepository;
-import kr.hhplus.be.support.exception.CommerceConflictException;
-import kr.hhplus.be.support.exception.ErrorCode;
+import kr.hhplus.be.domain.exception.CommerceConflictException;
+import kr.hhplus.be.domain.exception.ErrorCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
